+++ sema4ai/vscode-client/src/robocorpCommands.ts
@@ -27,6 +27,7 @@ export const SEMA4AI_ROBOTS_VIEW_TASK_DEBUG = "sema4ai.robotsViewTaskDebug";  //
 export const SEMA4AI_ROBOTS_VIEW_ACTION_RUN = "sema4ai.robotsViewActionRun";  // Launch Action
 export const SEMA4AI_ROBOTS_VIEW_ACTION_DEBUG = "sema4ai.robotsViewActionDebug";  // Debug Action
 export const SEMA4AI_ROBOTS_VIEW_ACTION_EDIT_INPUT = "sema4ai.robotsViewActionEditInput";  // Configure Action Input
+export const SEMA4AI_CONFIGURE_ACTION_INPUT = "sema4ai.configureActionInput";  // Configure Action Input
 export const SEMA4AI_ROBOTS_VIEW_ACTION_OPEN = "sema4ai.robotsViewActionOpen";  // Open Action
 export const SEMA4AI_RUN_ROBOCORPS_PYTHON_TASK = "sema4ai.runRobocorpsPythonTask";  // Run Sema4.ai's Python Task
 export const SEMA4AI_DEBUG_ROBOCORPS_PYTHON_TASK = "sema4ai.debugRobocorpsPythonTask";  // Debug Sema4.ai's Python Task