+++ docs/changelog.md
@@ -2,6 +2,8 @@
 
 ## New in 2.7.2 (2024-11-12)
 
+- Listing actions now will include `query` and `predict` decorators.
+- Add `Configure Action Input` internal command.
 - Set `SEMA4AI_CREDENTIAL_API` environment variable if port is communicated from Studio
 - Add packages and tools information when submitting an issue.
 - Update Action Server to `2.0.0`.